pyproject.toml
[build-system]
requires = ["hatchling", ]
build-backend = "hatchling.build"

[project]
name = "sparksampling"
description = "pyspark-sampling"
keywords = ["sparksampling", "pyspark-sampling"]
requires-python = ">=3.7"
dependencies = [
    "sparksampling-proto",
    'pyspark',
    'findspark',
    'traitlets',
    'pandas>=1.2',
    'requests',
    'kubernetes',
    'boto3',
    'grpcio-tools',
    'graphlib_backport',
]
dynamic = ["version", ]
classifiers = [
    'Development Status :: 2 - Pre-Alpha',
    'Intended Audience :: Developers',
    'Natural Language :: English',
    'Programming Language :: Python :: 3',
    'Programming Language :: Python :: 3.5',
    'Programming Language :: Python :: 3.6',
    'Programming Language :: Python :: 3.7',
    'Programming Language :: Python :: 3.8',
]

[project.scripts]
sparksampling = "sparksampling.app:main"


[project.optional-dependencies]
test = [
    "pytest",
    "pytest-grpc",
    "pre-commit",
]

[[project.authors]]
name = "Wh1isper"
email = "9573586@qq.com"

[project.readme]
file = "README.md"
content-type = "text/markdown"

[project.license]
text = "Apache License 2.0"

[project.urls]
Source = "https://github.com/Wh1isper/pyspark-sampling"

[tool.check-manifest]
ignore = [".*", ]

[tool.hatch.version]
path = "sparksampling/__init__.py"
